fix: resolve linting issues in test files
- Fix errcheck: properly handle Body.Close() return values
- Fix goconst: use constant for repeated "nonexistent" string
- Fix ineffassign: remove unused variable assignments
- Fix nlreturn: add blank lines before return statements
- Fix revive: rename unused parameters to _
- Fix unconvert: remove unnecessary int64 conversions
- Fix wastedassign: remove redundant stat assignments

+++ internal/web/template_cache_fuzz_test.go
@@ -34,8 +34,8 @@ func FuzzTemplateCacheSet(f *testing.F) {
 		if ttlMs < 0 {
 			ttlMs = 0
 		}
-		if ttlMs > int64(time.Hour.Milliseconds()) {
-			ttlMs = int64(time.Hour.Milliseconds())
+		if ttlMs > time.Hour.Milliseconds() {
+			ttlMs = time.Hour.Milliseconds()
 		}
 		ttl := time.Duration(ttlMs) * time.Millisecond
